Unifying som.cpp and lsat.cpp. Fixup unit tests for name change from genericsom -> som.
