+++ src/pj_list.h
@@ -55,7 +55,6 @@ PROJ_HEAD(geoc, "Geocentric Latitude")
 PROJ_HEAD(geocent, "Geocentric")
 PROJ_HEAD(geogoffset, "Geographic Offset")
 PROJ_HEAD(geos, "Geostationary Satellite View")
-PROJ_HEAD(genericsom, "Generic Space Oblique Mercator")
 PROJ_HEAD(gins8, "Ginsburg VIII (TsNIIGAiK)")
 PROJ_HEAD(gn_sinu, "General Sinusoidal Series")
 PROJ_HEAD(gnom, "Gnomonic")
@@ -148,6 +147,7 @@ PROJ_HEAD(s2, "S2")
 PROJ_HEAD(sch, "Spherical Cross-track Height")
 PROJ_HEAD(set, "Set coordinate value")
 PROJ_HEAD(sinu, "Sinusoidal (Sanson-Flamsteed)")
+PROJ_HEAD(som, "Space Oblique Mercator")
 PROJ_HEAD(somerc, "Swiss. Obl. Mercator")
 PROJ_HEAD(stere, "Stereographic")
 PROJ_HEAD(sterea, "Oblique Stereographic Alternative")
