+++ test/gie/builtins.gie
@@ -3689,7 +3689,7 @@ expect  127.761566096 -0.001716231
 ===============================================================================
 
 -------------------------------------------------------------------------------
-operation +proj=genericsom   +ellps=GRS80 +inc_angle=1.7157253262878522 +ps_rev=0.06866666666666667 +asc_lon=2.2298420007209447
+operation +proj=som   +ellps=GRS80 +inc_angle=1.7157253262878522 +ps_rev=0.06866666666666667 +asc_lon=2.2298420007209447
 -------------------------------------------------------------------------------
 tolerance 0.1 mm
 accept  2 1
@@ -3712,7 +3712,7 @@ accept  -200 -100
 expect  127.761567257 -0.001735150
 
 -------------------------------------------------------------------------------
-operation +proj=genericsom   +R=6400000 +inc_angle=1.7157253262878522 +ps_rev=0.06866666666666667 +asc_lon=2.2298420007209447
+operation +proj=som   +R=6400000 +inc_angle=1.7157253262878522 +ps_rev=0.06866666666666667 +asc_lon=2.2298420007209447
 -------------------------------------------------------------------------------
 tolerance 0.1 mm
 accept  2 1
